Old_development_docs/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
imblearn>=0.0
shap>=0.44.0
google-generativeai>=0.3.0
joblib>=1.3.0
matplotlib>=3.7.0
google-cloud-bigquery>=3.0.0
pyarrow>=10.0.0
statsmodels>=0.14.0
